services/Service_providers_service/requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
Django==5.0.1
django-js-asset==2.2.0
django-mptt==0.12.0
djangorestframework==3.14.0
djangorestframework-recursive==0.1.2
djangorestframework-simplejwt==5.3.1
idna==3.6
pillow==10.2.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
pytz==2023.3.post1
requests==2.31.0
rest-framework-simplejwt==0.0.2
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
